requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
spacy==3.7.2
python-dotenv==1.0.0
openai==1.3.7
jinja2==3.1.2
aiofiles==24.1.0